adding chart to app
